MODIFIED: Repo Layer Added(MySQL Database Connection)

+++ src/main/java/com/infogalaxy/librarymanagementsystem/entity/MemberEntity.java
@@ -14,7 +14,7 @@ public class MemberEntity {
 
     private String membername;
 
-    private int membermbno;
+    private double membermbno;
 
     private String memberemail;
